Changes after review: - coftest.py Class WebdriverFactory moved to helpers/webdriver.py - main_page.py: changes method is_ingredients_details_displayed() - test_order_feed_page.py: remove test transition
